feat: tech enhancement for adConfiguration

+++ PrebidMobile/PrebidMobileRendering/AdTypes/AdView/Modals/PBMModalViewController.m
@@ -14,7 +14,6 @@
  */
 
 
-#import "PBMAdConfiguration.h"
 #import "PBMClickthroughBrowserView.h"
 #import "PBMCreativeModel.h"
 #import "PBMMacros.h"
@@ -244,7 +243,7 @@
         PBMWebView *webView = (PBMWebView *)self.modalState.view;
         self.closeButtonDecorator.isMRAID = webView.isMRAID;        
     }    
-    self.closeButtonDecorator.buttonArea = self.modalState.adConfiguration.closeButtonArea.doubleValue;
+    self.closeButtonDecorator.buttonArea = self.modalState.adConfiguration.closeButtonArea;
     self.closeButtonDecorator.buttonPosition = self.modalState.adConfiguration.closeButtonPosition;
     [self.closeButtonDecorator setImage:[self.displayProperties getCloseButtonImage]];
     [self.closeButtonDecorator addButtonTo:self.view displayView:self.displayView];
