+++ PrebidMobile/PrebidMobileRendering/AdTypes/AdView/PBMVideoView.m
@@ -16,7 +16,6 @@
 #import <UIKit/UIKit.h>
 #import <MobileCoreServices/MobileCoreServices.h>
 
-#import "PBMAdConfiguration.h"
 #import "PBMCreativeModel.h"
 #import "PBMError.h"
 #import "PBMEventManager.h"
@@ -455,7 +454,7 @@ static CGSize const MUTE_BUTTON_SIZE = { 24, 24 };
 - (void)setupSkipButton {
     self.skipButtonDecorator = [AdViewButtonDecorator new];
     self.skipButtonDecorator.button.hidden = YES;
-    self.skipButtonDecorator.buttonArea = self.creative.creativeModel.adConfiguration.skipButtonArea.doubleValue;
+    self.skipButtonDecorator.buttonArea = self.creative.creativeModel.adConfiguration.skipButtonArea;
     self.skipButtonDecorator.buttonPosition = self.creative.creativeModel.adConfiguration.skipButtonPosition;
     
     UIImage *skipButtonImage = [UIImage imageNamed:@"PBM_skipButton" inBundle:[PBMFunctions bundleForSDK] compatibleWithTraitCollection:nil];
@@ -483,7 +482,7 @@ static CGSize const MUTE_BUTTON_SIZE = { 24, 24 };
         return;
     }
     
-    dispatch_time_t dispatchTime = [PBMFunctions dispatchTimeAfterTimeInterval:self.creative.creativeModel.adConfiguration.skipDelay.doubleValue];
+    dispatch_time_t dispatchTime = [PBMFunctions dispatchTimeAfterTimeInterval:self.creative.creativeModel.adConfiguration.skipDelay];
     @weakify(self);
     dispatch_after(dispatchTime, dispatch_get_main_queue(), ^{
         @strongify(self);
@@ -611,7 +610,7 @@ static CGSize const MUTE_BUTTON_SIZE = { 24, 24 };
 
     [self.avPlayer play];
     
-    [self handleSkipDelay:self.creative.creativeModel.adConfiguration.skipDelay.doubleValue videoDuration:self.creative.creativeModel.displayDurationInSeconds.doubleValue];
+    [self handleSkipDelay:self.creative.creativeModel.adConfiguration.skipDelay videoDuration:self.creative.creativeModel.displayDurationInSeconds.doubleValue];
 
     if (!self.isPlaybackStarted) {
         self.isPlaybackStarted = YES;
